move LinearTimeInplaceDFSRunner to a separate file
